+++ packages/internal/package.json
@@ -23,12 +23,11 @@
   },
   "dependencies": {
     "picocolors": "^1.0.0",
-    "sqids": "^0.3.0"
+    "workspace-tools": "^0.36.4"
   },
   "devDependencies": {
     "@aws-sdk/client-s3": "^3.441.0",
     "@types/node": "^20.9.4",
-    "vitest": "^1.6.0",
-    "workspace-tools": "^0.36.3"
+    "vitest": "^1.6.0"
   }
 }
