+++ packages/internal/src/getNextUpdate.tsx
@@ -1,0 +1,3 @@
+export const getNextUpdate = (appVersion: string) => {
+  return {};
+};
