+++ packages/internal/src/types.ts
@@ -9,3 +9,18 @@ export interface PluginArgs {
   secretKey: string;
   targetVersion?: string;
 }
+
+export interface UpdateSource {
+  [appVersion: string]: {
+    bundleVersion: number;
+    forceUpdate: boolean;
+    enabled: boolean;
+    files: string[];
+  }[];
+}
+
+export type UpdateSourceArg =
+  | string
+  | UpdateSource
+  | (() => Promise<UpdateSource>)
+  | (() => UpdateSource);
