+++ packages/internal/tsconfig.build.json
@@ -11,5 +11,5 @@
     "outDir": "lib"
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "src/**/*.test.ts", "src/**/*.spec.ts"]
+  "exclude": ["node_modules", "src/**/*.test.ts"]
 }
